+++ cli_tools/common/utils/logging/service/literal_loggable_test.go
@@ -1,0 +1,48 @@
+//  Copyright 2020 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package service
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestLiteralLoggable_GetValueAsInt64Slice(t *testing.T) {
+	loggable := literalLoggable{
+		int64s: map[string][]int64{
+			"gb": {1, 2, 3},
+		},
+	}
+
+	assert.Equal(t, []int64{1, 2, 3}, loggable.GetValueAsInt64Slice("gb"))
+	assert.Empty(t, loggable.GetValueAsInt64Slice("not-there"))
+}
+
+func TestLiteralLoggable_GetValue(t *testing.T) {
+	loggable := literalLoggable{
+		strings: map[string]string{"hello": "world"},
+	}
+
+	assert.Equal(t, "world", loggable.GetValue("hello"))
+	assert.Empty(t, loggable.GetValue("not-there"))
+}
+
+func TestLiteralLoggable_ReadSerialPortLogs(t *testing.T) {
+	loggable := literalLoggable{
+		serials: []string{"log-a", "log-b"},
+	}
+
+	assert.Equal(t, []string{"log-a", "log-b"}, loggable.ReadSerialPortLogs())
+}
